[revive] Receipts should include failed tx (#10120)
The current implementation incorrectly accumulates receipts in storage
to compute the `receipts_root`.
However, when an extrinsic fails, all storage changes are reverted
causing the receipts to be lost.

This PR updates the handling of Ethereum transactions so that they
always succeed. The transaction logic is now wrapped in a top-level
storage transaction that rolls back all state changes on failure while
still producing a valid receipt.

A new event `EthExtrinsicRevert` is emitted when an eth transaction
reverts.
eth-rpc uses this event to set the correct receipt status.

This event also serves as a replacement for ExtrinsicFailed to make
debugging failed Ethereum transactions easier in tools like Polkadot.js.

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>