Return the correct block difficulty from the eth-rpc (#10186)
# Description

This PR fixes an issue in the eth-rpc/pallet-revive that was causing it
to return an incorrect value for the block's difficulty or prevrandao.

In the VM/interpreter implementation we use a constant for the block
difficulty. However, the eth block construction side was unaware of this
constant being used and therefore the RPC was always returning a block
difficulty of zero.

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>