pallet_revive: use real storage when building the eth genesis block (#10225)
When building the eth genesis block, query the real storage items that
store the timestamp and block number instead of using the zeroed items.

If the chainspec does not customise these, they will remain zeroed.

However, anvil does customise these if requested by the user.
Needed in order to merge
https://github.com/paritytech/foundry-polkadot/pull/391