pallet-assets: add reserves info to support non-teleportable Foreign Assets on Asset Hub (#9948)
# Context

Asset Hub currently supports registering following types of assets:
1. Local Assets (also called trust-based assets), with `u32` asset ID,
asset creatable by Signed accounts and managed by `PalletInstance(50)`
(`Assets`).
2. Foreign Assets, `Location` as asset ID, creatable only by external
locations over XCM and managed by `PalletInstance(53)`
(`ForeignAssets`).

For "foreign assets", the current code assumes and enforces that the
asset is teleportable with its original location/creator.

Some ecosystem parachains would want to register foreign assets on Asset
Hub, but not also have them teleportable, not have Asset Hub act as a
reserve for the asset. Multiple reserve locations adds reserve
management overhead for teams, and some don't want to do it.

E.g. for HOLLAR to be registered to Asset Hub, Asset Hub would have to
allow reserve-based foreign assets.

# Solution

This PR adds a new storage item and new call to `pallet-assets` that can
be used to configure the trusted reserves (as seen by the local chain)
on a per-asset basis. This is a generic type that is flexible at the
runtime level. It can be an XCM Location or any other type of location
identifier, or it can be set to `()` if the assets managed by the pallet
instance do not require any remote reserves information.

The PR also adds runtime machinery to use the assets reserves
information when reasoning about XCM teleports and XCM reserve-based
transfers.

## AH considerations

In Asset Hub's case, for each asset, only explicitly configured assets'
reserve locations will be trusted. Besides the actual reserve
location(s), a`teleportable: bool` flag is also set for each reserve,
specifying whether the asset is teleportable or not between the local
chain and said reserve location.
Any other chain that should be assumed as a trusted reserve needs to be
marked as such in `pallet-assets` by the asset's `Owner`.

By default, on asset creation, no explicit reserve information is
configured. This results in foreign assets not being transferable
cross-chain by default. To enable cross-chain transfers, one of the
following configurations is required on the `Asset` to be done by its
`Owner`:
- For AH to accept incoming reserve transfer of asset "A" from remote
chain "X", that chain needs to be added as a trusted reserve for asset
"A", and `teleportable` flag set to `false`:
  ```rust
      let asset_id = "A";
      let origin = Signed(OwnerOfA);
      let reserves = vec![
ForeignAssetReserveData { reserve: "Remote Chain X location", teleport:
false }
      ];
      pallet_assets::set_reserves(origin, asset_id, reserves);
  ```
- For AH to accept teleports of asset "A" from remote chain "X", that
chain needs to be added as trusted reserves for asset "A", and
`teleportable` flag set to `true`:
  ```rust
      let asset_id = "A";
      let origin = Signed(OwnerOfA);
      let reserves = vec![
ForeignAssetReserveData { reserve: "Remote Chain X location", teleport:
true }
      ];
      pallet_assets::set_reserves(origin, asset_id, reserves);
  ```

## Required migrations

Since there are already a number of Foreign Assets already registered,
and have been teleportable or reserve-transferable since registration
based on static rules, this PR also adds a migration that configures the
required reserves information for existing foreign assets, so as not to
change any existing behaviors.

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Francisco Aguirre <franciscoaguirreperez@gmail.com>
Co-authored-by: Branislav Kontur <bkontur@gmail.com>
Co-authored-by: claravanstaden <claravanstaden64@gmail.com>