Enable force debug in revive dev node (#10383)
This change ensures that all types that implement `RuntimeDebug` are
fully displayed in log output of the revive dev node, instead of just
showing `<wasm:stripped>`.

Unfortunately, the trait `RuntimeDebugNoBound`, that we also use
frequently in pallet-revive, is not affected and will still output
`<wasm:stripped>` (it does not check for the `force-debug` feature flag,
instead it only fully outputs values when either one of the features
`std` or `try_runtime` is enabled – this is something we implement as a
general change).

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: pgherveou <pgherveou@gmail.com>