wasm-builder: Only overwrite wasm files if they changed (#10448)
When running two different `cargo` commands, they may both compile the
same wasm files. When the second `cargo` command produces the same wasm
files, we are now not gonna overwrite it. This has the advantage that we
can run the first command again without it trying to recompile the
project. Right now it would lead to the wasm files always getting
recreated, which is wasting a lot of time :)

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Sebastian Kunert <skunert49@gmail.com>